// apps/curator-web/src/api/generated/schema.d.ts
export type components = {
    schemas: {
// ...
        SessionResponse: {
            /**
             * Authenticated At
             * Format: date-time
             */
            authenticated_at: string;
            /**
             * Organization Id
             * Format: uuid
             */
            organization_id: string;
            /** Organization Roles */
            organization_roles: components["schemas"]["Role"][];
            /** Projects */
            projects: components["schemas"]["SessionProject"][];
            /**
             * Subject Id
             * Format: uuid
             */
            subject_id: string;
            subject_kind: components["schemas"]["PrincipalKind"];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
};